Psilocybin E-Coli
A small project i've been working on. I really wanted to share this, and update it as it progresses. The current plasmids do function accurately. Between 150ml/L to 1g/L depending on growth medium and mutation. I have used the T7 H10 promoter, and have plans to apply the site-directed mutagenesis to form better libraries.